All build targets are now under a single project. This gets rid of the need for different build directories to account for both varkor and the tests. They are all acounted for in the same build directory by using different targets. A msvc compiler flag to prevent implicit type conversions has been added. The readmes describing how to build the various targets have also been updated.

+++ src/math/Matrix.hh
@@ -64,7 +64,7 @@ std::ostream& operator<<(std::ostream& os, const Matrix<T, N>& matrix)
     {
       std::stringstream ss;
       ss << matrix[r][c];
-      int width = ss.str().size();
+      int width = (int)ss.str().size();
       if (width > longestWidths[c])
       {
         longestWidths[c] = width;
